move AnalyticsEventComponent to Analytics namespace
