+++ app/components/analytics/event_component.html.erb
